# docs/reference/audittrace/openapi.yaml
...
components:
  schemas:
...
    Body_upload_memory_file_memory_upload_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
      required:
      - file
      title: Body_upload_memory_file_memory_upload_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MemoryLayer:
      description: Valid memory layer targets for file upload.
      enum:
      - episodic
      - procedural
      - session
      title: MemoryLayer
      type: string
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    audittrace-oauth2:
      flows:
        authorizationCode:
          authorizationUrl: /protocol/openid-connect/auth
          scopes:
            audittrace:admin: 'Operator-grade administration: /memory/upload, /memory/index, hard-delete, configuration introspection.
              Optional scope on user-facing clients; never granted by default.'
            audittrace:assessment:ingest: Record a security self-assessment as audit events (ADR-058 recursive self-audit);
              a write grant distinct from the broad :audit read scope
            audittrace:audit: Read interactions, sessions, and tool-call audit rows
            audittrace:context: Read context-builder output for a query
            audittrace:index: Trigger semantic-store reindex (legacy dev client)
            audittrace:query: Issue chat completions and inference queries
            audittrace:scan:retrigger: Force a re-scan of an object in the ingestion content-control pipeline (ADR-048 operator
              surface). Admin-grade; never granted by default.
            memory:acl:read-own: Read effective permissions, accessible/public/sole-owned resource ids from the console-ACL
              store (Sovereign Authorization Layer EPIC, WU-1 — READ PATH ONLY) — the RLS-isolated store backing LibreChat's
              AclEntry records. Every answer is scoped to the CALLER's own resolved principal set (self + public); group principals
              are disabled at the schema level. This scope grants no write path (see memory:acl:write) and no access to any
              other memory layer.
            memory:acl:write: 'Write ACL grants through the console-ACL routes (Sovereign Authorization Layer EPIC, WU-2c):
              grant, revoke, modify, bulk and expire-by-predicate on the CALLER''s own resources. The database (RLS, migration
              032) decides who may write — a non-owner is refused and the refusal is audited. Does not open the read routes
              (memory:acl:read-own) and grants no access to any other memory layer.'
            memory:agents:read-own: 'Read your own agents (and batch-get-by-ids) from the console-agents store (Agents domain,
              MongoDB-elimination EPIC) — the RLS-isolated store backing LibreChat''s Agent records. Own-agents-only: sharing/marketplace
              is out of scope. Grants no access to another user''s agents or to any other memory layer.'
            memory:agents:write: Create/update/delete your own agents in the console-agents store (Agents domain, MongoDB-elimination
              EPIC). Grants no access to another user's agents (RLS-enforced) and no access to any other memory layer.
            memory:chat_projects:read-own: Read your own first-class chat-projects from the console-chat-projects store (Chat-Projects
              domain, MongoDB-elimination EPIC) — the RLS-isolated store backing LibreChat's chat-project groupings of conversations.
              Grants no access to another user's chat-projects or to any other memory layer.
            memory:chat_projects:write: Create/update/delete your own chat-projects in the console-chat-projects store (Chat-Projects
              domain, MongoDB-elimination EPIC). Grants no access to another user's chat-projects (RLS-enforced) and no access
              to any other memory layer.
            memory:conversation_tags:read-own: 'Read your own conversation-tags from the console-conversation-tags store (Conversation-Tags
              domain, MongoDB-elimination EPIC) — the RLS-isolated store backing LibreChat''s conversationTag records. Own-tags-only:
              sharing/marketplace is out of scope. Grants no access to another user''s conversation-tags or to any other memory
              layer.'
            memory:conversation_tags:write: Create/update/delete your own conversation-tags in the console-conversation-tags
...
info:
  description: 'EU-AI-Act-aligned audit trail for LLM workloads. OpenAI-compatible ``/v1/chat/completions`` proxy over a five-layer
    memory model (ADR-062): four per-user-isolated layers — episodic, procedural, conversational, semantic — plus Layer 5,
    the Shared Corpus (org-wide knowledge, shared-read, operator/curator-tier write). Keycloak-delegated identity, Postgres
    RLS per-user isolation, opt-in Redis-Streams async persistence (ADR-046), and a first-class audit event on every memory
    read/list/write/delete (ADR-062 §5 — ``event_class=memory_access`` on ``GET /interactions``).


    Recall (the LLM reading its own memory) happens through the ``recall_*`` tools inside the ``/v1/chat/completions`` tool-call
    loop (ADR-025) and through ``POST /context`` (the legacy ``inject``-mode aggregator) — NOT through the ``memory`` tag''s
    backoffice CRUD endpoints, which are the operator/audit surface for managing memory content directly.


    See ``docs/architecture/`` for C4 + sequence diagrams; the ADR set (014–062) is the source of truth for design decisions.'
  title: AuditTrace-AI
openapi: 3.1.0
paths:
...
  /memory/upload:
    post:
      description: "Upload a file to MinIO under the specified memory layer.\n\nTwo paths (Luis 2026-05-10, ADR-048 PR-B3):\n\
        \n* **PDF uploads** (``content-type: application/pdf`` AND\n  magic-byte sniff matches ``%PDF-``) take the quarantine\
        \ flow:\n  bytes land in ``quarantine/<user>/<scan_id>/<file>`` (not the\n  requested ``layer`` prefix), a ``memory_items``\
        \ row is inserted\n  with ``scan_status='pending_scan'``, and an AMQP scan-request is\n  enqueued. The response is\
        \ **HTTP 202** with a ``scan_id`` and\n  poll URL. Promotion to ``episodic/papers/`` happens\n  asynchronously in\
        \ content-control once the verdict is clean. This\n  branch is UNCHANGED by ADR-062 Phase B — the async quarantine\
        \ →\n  verdict → promotion pipeline is ADR-048's content-control\n  concern, out of WU-B4/WU-B5's scope.\n\n* **Non-PDF\
        \ uploads** (markdown skills, ADRs, plain text): ADR-062\n  Phase B (WU-B5) — writes to the caller's PRIVATE tier\
        \ via the\n  episodic/procedural service's own ``write()`` (already wired to\n  the private bucket by WU-B2), closing\
        \ a leak where every non-PDF\n  upload used to land unconditionally in the shared/corpus bucket\n  regardless of who\
        \ uploaded it. Neither layer has a declared\n  corpus-write scope (ADR-062 §4's frozenset is exactly\n  decisions/skills/semantic),\
        \ so ``?promote=corpus`` is rejected\n  with 400 rather than silently accepted or silently ignored.\n\n* **``layer=session``\
        \ uploads** (WU-1, Sovereign-Attach EPIC): a\n  THIRD case of the non-PDF branch — writes to the Postgres-backed,\n\
        \  per-user, EPHEMERAL session layer instead of S3\n  (``services/session_memory.py``). PDF content is refused for\
        \ this\n  layer (400) rather than routed into the quarantine pipeline above:\n  that pipeline promotes a clean verdict\
        \ into ``episodic/papers/``\n  (DURABLE) regardless of the ``layer`` query param — silently\n  letting it through\
        \ would let a token holding ONLY\n  ``memory:session:write`` (nothing durable) land content durably\n  anyway, defeating\
        \ this WU's least-privilege wall.\n\nAuthorization (per-layer): the caller's JWT must carry\n``memory:<layer>:write``\
        \ matching the ``layer`` query parameter\n(or ``audittrace:admin``). A token with ``memory:procedural:write``\ncannot\
        \ upload to ``layer=episodic`` and vice-versa; a token with\nONLY ``memory:session:write`` cannot upload to any durable\
        \ layer.\nThis is enforced by ``_require_layer_write`` on the first line below\n— the endpoint is **not** auth-only\
        \ despite the static ``scopes=[]``\non the OAuth2 declaration (which exists only so the spec lists the\nsecurity scheme;\
        \ the effective scope is dynamic in the ``layer``\nparameter, so it cannot be declared statically — same contract\
        \ as\n``/memory/index``)."
      operationId: upload_memory_file_memory_upload_post
      parameters:
      - in: query
        name: layer
        required: true
        schema:
          $ref: '#/components/schemas/MemoryLayer'
      - in: query
        name: filename
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
      - description: Set to 'corpus' to request a corpus-tier write. Neither episodic nor procedural has a corpus-write scope
          declared (ADR-062 §4) — any corpus request here returns 400.
        in: query
        name: promote
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Set to 'corpus' to request a corpus-tier write. Neither episodic nor procedural has a corpus-write
            scope declared (ADR-062 §4) — any corpus request here returns 400.
          title: Promote
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_memory_file_memory_upload_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Upload Memory File Memory Upload Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2:
        - memory:episodic:write
      - audittrace-oauth2:
        - memory:procedural:write
      - audittrace-oauth2:
        - audittrace:admin
      summary: Upload Memory File
      tags:
      - memory
  /memory/upload/status:
    get:
      description: 'Return the current scan_status of the named upload.


        Authorization: any authenticated user who can read the

        layer can poll status. We additionally check that the

        requesting user is the same one who created the row OR an

        admin — content-control verdicts shouldn''t leak across

        tenant boundaries even pre-PR-B7 IAM split.'
      operationId: get_upload_status_memory_upload_status_get
      parameters:
      - description: Scan ID returned by POST /memory/upload
        in: query
        name: scan_id
        required: true
        schema:
          description: Scan ID returned by POST /memory/upload
          title: Scan Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Upload Status Memory Upload Status Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - audittrace-oauth2: []
      summary: Get Upload Status
      tags:
      - memory
      - memory
...